Introduced threaded processing
Made an utility class for processing in a separate thread. This is
desirable, as then the main GUI thread is freed for processing the event
loop. In other words, it keeps the application from becoming
unresponsive. A progress bar pops up, which can either just show "busy"
state, or percentage of completion. Some care should be taken when
processing in a separate thread, as drawing is only allowed on the main
thread. If needed, this should be handled by events.