Uses traitsui to edit component parameters.
By changing the Hyperspy code (thereore the branch) to make all
components and parameters traits.api.HasTraits, the components can be
edited by a traitsui dialog. Next step is to capture the dialog and add
it to a permanent QWidget panel at the bottom of the tree (visible when
component selected).