Remove mem leak
The standard MPL Qt backend leaks objects. Our backend implementation
was in large parts copied from that, so we were leaking as well, which
meant that Signals that had been plotted were never freed. This update
fixes that, although it was mostly solved by trial and error, so things
could pop up later.