Plugin (un/re)loading on the fly
Added facilities to load, unload or reload plugins on the fly. This was
mainly achieved by adding methods on PluginManager, but Plugin also had
to get an unload method. A default unload method was specified, which
means no changes have to be done in the implemented plugins, except to
call local wrappers for MainWindow API.