Improved plugin encapsulation
Now all calls to plugins from the PluginManager are wrapped in a try
block that down-converts any non-system exceptions to warnings. Any
other calls into the plugins should be during program execution (as
opposed to loading), so they will "safely" unwind to the event loop, and
not crash the application.